Many fixes to bible.csv
Fixed numerous locations where the CSV formatting was not done properly,
resulting in unformatted Bible verses.

There were also some typos which I managed to spot and fixed, referring
to a digital scan of the New Jerusalem Bible for clarity on parts where
the formatting was a bit dodgy.

There were some lines where I noticed words and spaces were clearly
missing, so I referred to a scan of the NJB to amend those (manually).

(By the way, this is the New Jerusalem Bible, not the Jerusalem Bible)

Additionally, I added almost all missing verses, referring meticulously
to the footnotes of the New Jerusalem Bible (admittedly, I may have
added them a bit too liberally)
They are all indicated with a * for clarity (and so users know it was
not in the plain text of the NJB but in the footnotes.